backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
pillow>=10.4.0
torch>=2.2.0
torchvision>=0.17.0
ultralytics>=8.2.0
onnxruntime>=1.18.0
numpy>=1.26.0
scikit-learn>=1.5.0
matplotlib>=3.9.0
